fixing creation and deletion performance tests
